+++ bean-searcher-boot-starter/src/main/java/cn/zhxu/bs/boot/BeanSearcherProperties.java
@@ -667,11 +667,11 @@ public class BeanSearcherProperties {
 		private boolean useJson = true;
 
 		/**
-		 * 使用 {@link JsonFieldConvertor } 时，当遇到某些值 JSON 解析异常时，是否自动捕获（即忽略）
-		 * @see JsonFieldConvertor#setCatchError(boolean)
+		 * 使用 {@link JsonFieldConvertor } 时，当遇到某些值 JSON 解析异常时，是否抛出异常
+		 * @see JsonFieldConvertor#setFailOnError(boolean)
 		 * @since v4.0.1
 		 */
-		private boolean jsonCatchError = false;
+		private boolean jsonFailOnError = true;
 
 		/**
 		 * 是否启用 {@link ListFieldConvertor }，默认为 true
@@ -805,12 +805,12 @@ public class BeanSearcherProperties {
 			this.useJson = useJson;
 		}
 
-		public boolean isJsonCatchError() {
-			return jsonCatchError;
+		public boolean isJsonFailOnError() {
+			return jsonFailOnError;
 		}
 
-		public void setJsonCatchError(boolean jsonCatchError) {
-			this.jsonCatchError = jsonCatchError;
+		public void setJsonFailOnError(boolean jsonFailOnError) {
+			this.jsonFailOnError = jsonFailOnError;
 		}
 
 		public boolean isUseList() {